feat: move add-student to standalone page, add selector option
- /students/new: standalone page with form; redirects to /dashboard on success
- Student selector dropdown gains "+ Add a Student" option → navigates to /students/new
- Dashboard Students card shows list + link to /students/new; inline form removed
- dashboard/actions.ts trimmed to logoutAction only
